add urlquote wrapper in deriva/core/__init__.py with standards-safe defaults
The urllib.urlquote function has an unsafe default behavior.

We always want to encode **all** reserved characters in our REST API
elements, including `/` which is excluded for terrible legacy reasons
in the urllib version.  We should expect arbitrary user-generated
content to be properly quoted, which means every input reserved
character including `/` MUST be encoded.

If this breaks anything else in deriva, it is the **calling** site
which is broken, not this wrapper.