add deriva.core.ermrest_model.Table.define_vocabulary() method and docs
This method makes it trivial to create vocabulary tables:

    schema.create_table(
      Table.define_vocabulary(
        "My Vocabulary Table",
        "MYPREFIX:{RID}",
        "https://myserver.example.org/id/{RID}"
      )
    )

It is a convenience wrapper that calls Table.define() while generating
the core vocabulary column and key definitions.  It also accepts the
same arguments as Table.define() and passes those through so that a
caller can extend or further customize these core structures.

The curie_template and uri_template features depend on new domain types
and related support in the latest ermrest service code.