override requests.Response.raise_for_status to provide more detail
Our deriva server responses have textual content that provides more
error information to clients.  The requests.Response.raise_for_status
method makes the response object available on thrown HTTPError
exceptions but does not include the content in the exception text that
is displayed by default when an exception is not caught by naive
calling scripts.

This patch adds a local _response_raise_for_status(response) function
and calls it from each place we previously called
response.raise_for_status().  This adds an extra "details: ..."
suffix to the exception string including our response.content.

A bit of dead-code is also included in ermrest_model to show how to
monkey-patch this function on top of a response.raise_for_status
method.  It is dead because we already throw an exception, but if we
ever change the API to return response objects and expect the caller
to raise_for_status(), we would want to provide them with this
patched variant.