improve config equivalence to avoid some recurring updates to ACL bindings
The ACL binding equivalence check may have false negatives due to the
server filling in defaults so that the published binding is not
identical to the input binding document.  This cases repeated PUT
requests which are idempotent and wasteful.

We can help a common case where `scope_acl` is omitted and defaults to
the `["*"]` value. However, we cannot easily help a case where
`projection_type` is omitted, because determination of the default
depends on projected column type inference which we don't want to
bother doing in this client library...