changed the dataframe property to throw exception is pandas not present
previously, the Entities.dataframe property would return None if the
pandas library was not present. in the revised code, the api will
attempt to import pandas and therefore raise the ImportError if not
available. still the api allows use of datapath api (and therefore all
of deriva-py) without the user installing pandas, which should remain
optional. also, a test is added to the unit tests.