refactored datapath to encapsulate more of its internal state
Certain internal properties of various datapath objects had been
"exposed" in the sense that the property names were not prefixed with an
'_'. There was a history for why this was, but there's really no reason
for these to be confused for general usage.

Classes too have been marked as internal with leading '_' and some have
been renamed as "...Wrapper" to clearly indicate these are not intended
as replacements or extensions of the respective model classes.

Also, the 'describe' method has been deprecated and should be dropped in
a future release.