reintroduce some dynamic properties on column objects (issue #134)
In prior work, datapath was refactored to scale back on the use of
dynamic properties just to simplify the code base. This introduced a bug
in the naming of columns when they were bound to table alias instances.
URIs would therefore be generated with the original table name rather
than the table alias name, resulting in bad fetch URLs. Along with the
fix, a couple new tests were introduced to cover these cases.