also apply annotations and acls in final stage for pre-existing tables
The original cloning logic failed to account for applying configuration
changes to tables that already exist in the destination catalog. This
is usually just the built-in ERMrest tables in the public schema,
but could include others if the caller supplies the destination catalog
as an input argument.

To be safe, apply all configuration again in the final phase. The
normal apply() logic tries to avoid unnecessary requests and so should
be efficient. Also, we do not want to revert the cloning status
state value already written into the destination table annotations
by earlier clone processing stages, so that gets special handling.