datapath denormalize, fix logic to skip whole entities
fixes a bug that would have resulted in removing a column from the
projection if it was part of a fkey that referenced a table that did not
have a 'name' like column in it. now, if the 'name' is not found it will
at least project the referenced table's 'RID'. and renamed the option to
get whole entities (ie, project '*') with a positive 'include...' rather
than negative 'skip...' semantics.